Moved icons to shared folder
