+++ src/domain/collaboration/canvas/CanvasesDashboardSection/CanvasesDashboardSection.tsx
@@ -4,7 +4,7 @@ import DashboardGenericSection, {
 import { Button, Typography } from '@mui/material';
 import React from 'react';
 import { Add } from '@mui/icons-material';
-import { CanvasIcon } from '../../../../common/icons/CanvasIcon';
+import { CanvasIcon } from '../../../shared/icons/CanvasIcon';
 import { useTranslation } from 'react-i18next';
 import SimpleCard from '../../../shared/components/SimpleCard';
 import { LinkWithState } from '../../../shared/types/LinkWithState';
