+++ src/domain/platform/admin/templates/CanvasTemplates/CanvasTemplateCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { CanvasIcon } from '../../../../../common/icons/CanvasIcon';
+import { CanvasIcon } from '../../../../shared/icons/CanvasIcon';
 import SimpleCard, { SimpleCardProps } from '../../../../shared/components/SimpleCard';
 
 interface CanvasTemplateCardProps extends Omit<SimpleCardProps, 'iconComponent'> {}
