+++ src/domain/shared/icons/CanvasIcon.tsx
