+++ src/domain/shared/icons/ChallengeIcon.tsx
