+++ src/domain/shared/icons/HubIcon.tsx
