+++ src/domain/shared/icons/LibraryIcon.tsx
